+++ cxx/graphidx/tree/postorder.hpp
@@ -101,5 +101,5 @@ public:
 
     const int* end() const { return postord + (include_root ? n : n-1); }
     
-    size_t size() const { return end() - begin(); }
+    size_t size() const { return size_t(end() - begin()); }
 };
