+++ cxx/graphidx/tree/prufer.hpp
@@ -11,14 +11,14 @@
  */
 template<typename int_ = int>
 inline int_
-prufer2parent(const int_ n,
+prufer2parent(const size_t n,
               const int_ *prufer,
                     int_ *parent)
 {
-    for (int_ i = 0; i < n; i++)
+    for (size_t i = 0; i < n; i++)
         parent[i] = -1;
     std::vector<int_> degree (n, 1);
-    for (int_ i = 0; i < n-2; i++) {
+    for (size_t i = 0; i < n-2; i++) {
         const auto pi = prufer[i];
         if (pi < 0 || pi >= n)
             throw std::invalid_argument(std::string("prufer[") +
